WIP: deserialize the whole Block. Fixed deserialized InMsgDescr. Added AccountStates enum. Removed skipMagic.

+++ cell/src/main/java/org/ton/java/tlb/types/MsgAddressIntStd.java
@@ -25,7 +25,7 @@ public class MsgAddressIntStd implements MsgAddressInt {
     int magic;
     Anycast anycast;
     int addrLen;
-    int workchainId;
+    byte workchainId;
     BigInteger address;
 
     @Override
